Enable testing with firebase emulator.
Fixed the bug mentioned in previous commit where the app could not be deployed when using node.js imports for jest. The main part of this fix was implementing a way to feed the corrrect firebase and corresponding firestore instance to teh source code based on if the code was running in node.js or not.